+++ CHANGELOG.md
@@ -119,7 +119,7 @@
 - Always use Javascript conditional blocks (#333)
 - links in stats pane don't set up the correct date range (#106)
 - code should delete scroll id after use (#336)
-- ll.py - Make --count work with --pretty
+- ll.py - Make --count work with --pretty; show private message counts
 - DRY: move leapYear and end of month calculations to utils
 
 ## CHANGES in 0.9b: